Micro-optimizations in the decoder.
Mostly around reducing slice manipulation and integer conversion.
Benchmarks `read_decode` and `write_decode` show ~3% improvement.

GitOrigin-RevId: 1953b6abefa749929627edc55cc50dd88f4d1854

+++ fuse/server/decode_test.rs
@@ -39,7 +39,7 @@ fn request_decoder_new() {
 
 	assert_eq!(
 		decoder.consumed,
-		size_of::<fuse_kernel::fuse_in_header>() as u32
+		size_of::<fuse_kernel::fuse_in_header>()
 	);
 }
